move errors up
